tdf#150118: fix call to AccessCheck
It turns out, that despite PrivilegeSet argument is documented as
optional, the function fails without it. So my advise to avoid it
was incorrect after all.

Also, SAL_WARN takes area as the first argument, and a stream as
the second. It makes no sense to "recode" GetLastError result: it
is most informative as is.

Change-Id: Ifd4d4ddc96f5e001c898f7331148b47aeee77f2f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192227
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 1a4f72f03b1bd49d486e52135e78971242fae649)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192231
Tested-by: allotropia jenkins <jenkins@allotropia.de>
Reviewed-by: Thorsten Behrens <thorsten.behrens@collabora.com>