tdf#101630: Use redirect URL for browser-based OAuth2
To interact with an OAuth2 provider, we must open a browser with the
URL of the provider, and in that URL encode our target URL to use to
send information about the result of authentication (either failure,
or the access tokens).

The target URL can either be an HTTP URL, or a custom URI scheme [1].
The custom scheme looks attractive, because it doesn't need to listen
to HTTP, and would receive the data through command line arguments.
However, this technique requires to register a custom URI handler on
the system; and that can't be done temporarily, for the current app,
on some platforms; so if we decided to use the custom URI scheme, it
would only work for installed app, not for portable one.

Using HTTP URL, it can be either an external one (which would require
to run an online service for authentication, which would receive the
data from OAuth2 providers, and then communicate it to the app that
requested it - with all kinds of security implications), or it can be
a loopback URL to a local HTTP server. The latter option has no HTTPS
requirement (the sensitive data is sent through HTTP locally, never
leaving the host; the problem with HTTPS is that we would need a cert
for the local system, trusted by browsers); that is implemented here:
a local server is established by the app itself, before sending the
auth request.

This change only provides implementation for Windows. To have a cross-
platform solution, another HTTP server implementation is needed, e.g.
using boost/beast.

This also removes password prompt from the File Services dialog for
Google Drive and OneDrive, when the new method is implemented (i.e.,
currently only on Windows).

[1] https://www.oauth.com/oauth2-servers/oauth-native-apps/redirect-urls-for-native-apps/

Change-Id: I4684477c73a54370721e700fa95fc5f14344cf60
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192253
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>