tdf#168559 PPTX imp: fix missing list style for outline shapes on master pages
Load oox/qa/unit/data/outliner-list-style.odp into Impress, save as
PPTX, load it the PPTX, save the PPTX again (source is now PPTX), open
it again, go to the end of the first bullet, enter, tab, indent grows to
a large value.

What happens is that the PPTX export could write the correct list style
for the shape (based on the master page) when the source was ODP, but it
can't do the same when the source was PPTX, so after all this a PPTX
import problem.

Fix this by improving the formatting of the outliner shape on master
pages: if the outliner shape was empty, then we don't insert formatted
text body, so instead try to format the existing content of the outliner
shape, given that such shapes have pre-existing content on the master
page.

Notes:
- Do this only for outliner shapes, CppunitTest_sd_import_tests2's
  testTdf103792 is a sample that shows doing this for all placeholders
  on the master page is not correct.
- The first line offset is typically negative, but do guard for positive
  values, otherwise the exported ODP would be invalid, see
  CppunitTest_sd_export_tests's testTdf128550.
- It's possible to have outliner shapes with custom prompt text, which
  is typically not bulleted, so leave those alone, otherwise we would
  lose that prompt text, as visible in
  CppunitTest_sd_export_tests-ooxml4's testCustomPromptTexts.

Change-Id: I5b1768a4a4b9cb8494785a38e040dacdf4ab5289
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192270
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins