tdf#168406: oox: Restart numbering after non-numbered element
If we get a numbered list element, followed by a non-numbered element
then another numbered element it looks like pptx expects
the numbering to restart.

There are two halfs to this:

a) when we find ourselves in a numbered list (buAutoNum) we record
it in the recently added tracking mask, and if it turns out we
just entered the list trigger the numbering restart.

b) We have to detect when we get a paragraph which isn't numbered -
i.e. doesn't have buAutoNum (it turns out there are cases without
buNone), and when we detect that update the list tracking mask.

Change-Id: I9951147b3b7ec78009f2f87c93bbda9a17c0db28
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192414
Reviewed-by: David Gilbert <freedesktop@treblig.org>
Tested-by: Jenkins