tdf#168688 DOCX import: fix missing char props in field in header in ToC
Regression from commit 6ee33d0772f991f04e97b5b61d7b11810a1b8ac0
(tdf#104079 RTF import: fix handling fields inside TOC fields,
2017-10-17), open the bugdoc, check the font size of "TEST" in the
header, it should be 24pt, but it's 12pt.

What happens is that the document ends with the table of contents DOCX
field, so the section properties at the end of the document (including a
reference to a header) get parsed while the ToC field is still open, so
we think that m_xTOCMarkerCursor is in use, but it's for the body text
and we're in the header, so it's not relevant.

Fix the problem by moving m_xTOCMarkerCursor to m_StreamStateStack,
where all state that should be stashed away & restored later belongs.
This way the old tdf#104079 use-case keeps working and the new use-case
gets fixed.

Interestingly this xTOCMarkerCursor is not actually used as a cursor
anymore and it just indicates if we're inside a ToC or not, but leave
that unchanged in this commit.

Change-Id: I1d1c9f55413d94dcde9e7c715bc9f9863f354dcb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192392
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit 86d0a2f84d714da67d83fdba652540ebe1027382)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192720
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>