cool#13193 Fix Ctrl Shift V issue on impress
f0c25c751cf8e166a84b289746bce6202a40391d is merged to fix tdf#115783
But that commit caused a bug on Impress when the we have different
character and paragraph styles in textboxes. Commit aims to solve a
table issue.

To solve the Impress regression we should
apply the solution only table objects

I tried to write a test but it's not testable with reasonable
amount of effort

Signed-off-by: Gülşah Köse <gulsah.kose@collabora.com>
Change-Id: I132e8257af2aadbdb816865678966b54de373ef8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192389
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194947
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins