Related: tdf#168842 digits in GtkSpinButton, lost in SRC -> UI conversion
The digits setting may only apply when the current unit is point: see
SetFieldUnit, which special-handles the unit. So the omission of the
setting went unnoticed for a long time.

This is similar to commit 75b30f1d107f45aebbf9a0811b9d40f4b2bafdce
(tdf#168842 Issue entering paragraph values in cm, when default is in
points, 2025-10-15), and tried to handle the rest of omissions in UI.

Change-Id: Ia348e29b268c5f49dc9eedc24262d30efbb8faa5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192636
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins