Related: tdf#168842 Don't truncate source precision in vcl::TextToValue
When user enters a value like "2.54cm", and current input box is points,
which has only one decimal digit, it does not make sense to apply that
precision to the input string, making it mean "2.5cm" (and resulting in
70.9 pt).

This change adds a function to convert such strings to double, instead
of sal_Int64; then the double is handled using vcl::ConvertDoubleValue,
as before.

Now entering "2.54cm" produces the expected 72.0 pt.

This needed some OUString->u16string_view for loplugin:stringviewparam.

Change-Id: I5afa4edada904c4d8b8bcd9df65243546cb6bd00
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192663
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>