tdf#130857 qt: Extract helper to convert tooltip text to HTML
Extract existing logic to a helper function, for reuse
in other places handling tooltip texts in an upcoming commits.

Explicitly handle an empty string arg to also return
an empty string, as some QToolTip logic triggered with an upcoming
commit in place skips showing a tooltip when an empty string
is used, but converting to HTML would result in a non-empty
tooltip string, even if there's no actual text to show.

Change-Id: Ie863daff72be866731c3fc0d3618dd55bb0ad127
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192398
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>