tdf#101083 speed up SVG rendering with pattern fill (2)
This is the same thing I did in

    commit 3cbe3a0259bea4dec70e72191ec3c03441926a07
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Mon Jun 14 15:05:59 2021 +0200
    tdf#101083 speed up SVG rendering with pattern fill

but cut and paste from the VclPixelProcessor2D to the CairoPixelProcessor2D code.

Which makes this SVG render in reasonable time when using the cairo backend.

We can probably do even better by cutting out the usage of VCL here, but this
is good enough for now.

Change-Id: I959539741eef627b25702c91cad3e5f8ee2b3176
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192394
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>