tdf#130857 qt weld: Use HTML for tooltip texts
Instead of setting the plain text tooltip as is,
use HTML formatting.

This ensures that the tooltip text wraps.

This basically applies the same solution as introduced
for the non-weld case in

    commit fa89130d4980662642cb81a05ec044dbe21fd5d0
    Date:   Fri Aug 2 19:30:18 2024 +0200

        tdf#162297 Wrap tooltips in Qt5

        Acoording to the Qt documentation, in order to wrap the tooltip a rich-text style needs to be set.

for the case of native Qt widgets as well and
e.g. makes the extended tip as described in
tdf#162297 wrap for the SAL_VCL_QT_USE_WELDED_WIDGETS=1
case as well.
In the same way, the tooltip mentioned for Calc's
"Sort" dialog case in [1] also wraps with this commit
in place when using the qt6 VCL plugin with
SAL_VCL_QT_USE_WELDED_WIDGETS=1 after support
for that dialog was declared in

    Change-Id: Ica500005020169de0faf2dada052e5d8f5665ce9
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Oct 14 14:10:00 2025 +0200

        tdf#130857 qt weld: Support Calc "Sort" dialog

Also adjust QtInstanceWidget::get_tooltip_text
to do the reverse conversion from HTML to plain
text again, using QTextDocumentFragment that is mentioned
as a potential solution for such a task in [2].

[1] https://lists.freedesktop.org/archives/libreoffice/2025-October/093867.html
[2] https://stackoverflow.com/questions/2799379/is-there-an-easy-way-to-strip-html-from-a-qstring-in-qt#comment73105365_9546425

Change-Id: Ie57d59e482103b25a94de8377c0f5562ff294b46
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192399
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins