ui: enable single-click activation for writer navigator panel
context:
there is not special case handler exist for writer navigator on single click. it just highlight the entry on single click. It will be better if we enable the entry on single click. It will fixed the issue where pressing "enter" or "space" in writer navigator, doesn't move the view to the selected place in outline tab

Signed-off-by: Parth Raiyani <parth.raiyani@collabora.com>
Change-Id: I352a211cf21e6926ee3f5e83d2b0c91f9bd0dae3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192408
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Szymon Kłos <szymon.klos@collabora.com>