retain original FSType
This ends up as the OS/2 fsType field, retain the original font
FSType here.

what matter is:

- [makeotf] The OS/2 table 'fsType' field is now set to the feature file
  override. If not supplied, then the value of the environment variable
  FDK_FSTYPE. If not set then 4 (Preview & Print embedding).

so always put this in the features file on generating the otf, but while
we're at it retain the FSType in cidfontinfo description as well for
consistency.

This retains the original embedding permissions without defaulting to
bogus 'cannot edit' messages if the original fonts allowed editing.

Change-Id: I1c40a62ac9c0371543f7f599bee2069439293d8a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192449
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins