sc: delete sheet views when the table is deleted
This deletes all the sheet view holders when the table that has
any sheet views is deleted. If a sheet view holder table is
deleted, this change also makes sure that the sheet view gets
properly deleted.

This change also adds the test for deletion, that asserts step by
step that the values tables and sheet views are inserted and delete
correctly. Result of this is a couple of fixes to the delete
operation.

Change-Id: I29594c5e660fe8c42715b08bb9dcc3f25a611f29
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192423
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>