tdf#165615 sw floattable, no overlap: avoid check against an own follow
Load the bugdoc, the layuot loop control kicks in:
warn:sw.layout:20117:20117:sw/source/core/layout/tabfrm.cxx:1853: LoopControl in SwContentFrame::CalcLowers
and then the content is laid out on 3 pages, while it fits 2 pages in
Word.

What happens is that the document has an inline table, then a floating
table inside that. The inner floating table has tblOverlap set to
'never', and then things go wrong in
SwToContentAnchoredObjectPosition::CalcOverlap().

Fix the problem by checking if the two overlapping fly frames are
actually next to each other in a split fly chain: if we're the precede
of the other fly, then no need to worry about the overlap, that follow
fly will move to the next page anyway.

In theory this would be also interesting for toplevel floating tables,
too -- but in practice CppunitTest_sw_core_layout's testSplitFlyHeader
shows doing this in general would need more work, so just do this for
floating tables which are inside tables.

Change-Id: I4d0f1bd5229f8fbc3bb49af57f88fd6ef911cbfe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192426
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>