tdf#72341 sw: Make inc/dec font size affect all scripts
This change updates the FN_GROW_FONT_SIZE and FN_SHRINK_FONT_SIZE
handlers to scale fonts for all script types, even if the
currently-selected text does not contain a particular type.

This fixes the reported issue that bullets in RTL context are not
resized if a selected list item does not contain CTL characters. This
also fixes the broader issue of inserted text not reflecting earlier
size changes, simply because the script type for that inserted text
wasn't present at the time.

Change-Id: I881f0b90b5c3aecbbf2e40ee2ad6758de9c5bad3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192459
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192725
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>