tdf#168861 tooltip line wrap for length > 100 char
The current threshold is 150 characters. However, this results in very
wide normal tooltips that are hard to read. Therefore, the threshold is
reduced to 100 characters. The value 100 was chosen because it is
currently already used as threashold for cases where the text contains
line breaks.

If the tooltip length is smaller than 100 characters, the tooltip is
displayed in a single line. This behavior does not change with this
patch.

The threshold was introduced for OOo bug i42424 to force a line wrap
in the footnote balloon rectangle that is shown when you hover over
a footnote anchor with the mouse.

I don't know how to test the width of such a temporary tooltip
rectangle in a unit test. Thus a unit test is missing.

Change-Id: I7e4c47e5212f0af2510878ef2c3d6475d1d5ba46
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192458
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
(cherry picked from commit 34612217f8a027b940169899eb5c5514155649df)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192570