tdf#166044 sw: Fix cont. footnotes repeating the dest page number
Fixes an issue causing Writer to print the destination page number twice
in footnotes with continuation notices that end in non-Latin text.

This was happening due to subtle mismanagement of the m_bHasFollow flag.
Continuation notices rely on this flag being unset to indicate the field
portion responsible for printing the destination page number, but the
layout algorithm was leaving this flag unset on two different portions.
In Latin-only documents, this bug was hidden by SwLineLayout::CalcLine()
removing the second portion, but this isn't possible when a script
change is involved.

Regression since commit 6c0105307efa15837e9e5287103995312f373b67
(sw: fix stale SwNumberPortion follow flag)

Change-Id: I94c6f7ca75a3cf844d3eff2640d63777ac026717
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192464
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>