extension mgr: Use -1 to indicate non-selected entry
Instead of having two different members,
ExtensionBox_Impl::m_bHasActive and
ExtensionBox_Impl::m_nActive to track whether
an entry is currently active/selected, and if so,
which one, use only ExtensionBox_Impl::m_nActive
and assign a special value of -1 when no entry is
selected.

That also matches the ExtensionBox_Impl::getSelIndex
logic which was already returning -1 in case of
no entry being selected.

No change in behavior intended.

Change-Id: Iecdb7ce22680a42f100d5e71dc36432f037c65f7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192490
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins