tdf#168406: sd/qa: Add numbered list content check
Add a basic numbered list content check using a pptx import.
The file partially overlaps with tdf#168406 problem, but I've
not yet managed to find a way to create something quite as weird
as the original.

In sd/qa/unit/layout-tests because it's got most of the XML wrangling locally.

The test file was created using MS Onedrive/365 (free version).

Change-Id: Iee566c961ec5e01a7e2640eaf30839a0114fdcd2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192517
Reviewed-by: David Gilbert <freedesktop@treblig.org>
Tested-by: Jenkins