sc: Fix issues with tile invalidation
Include sheet view in a view render state, so we get the
correct tiles for our current view.

Also use CurrentTabForData to get the anything connected to the
size of columns and rows. We need to get that from the holder
sheet if we hide, show, change the size, of columns and rows.

Change-Id: I9431488d43920f5bb492c4bdcbc83dd24db0d89f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192568
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>