generictoolbarcontroller: Flatten the early-return checks in execute
Instead of having a bunch of nested if statements in order to initialise
xDispatch while the solar mutex is locked along with a check for whether
the initialise happened after the lock is released, there are now just
early returns on failures. This should make it a bit easier to add an
extra step for the success case while holding the mutex in a subsequent
patch.

Change-Id: I731a800598a8e31b2b18ee29b6354c5440a430dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192648
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>