tdf#130857 weld: Move logic for (not) calling signal hdls to base classes
Following previous commit

    Change-Id: I93ec71e29804fd601e314e24c96d0f18a7cbb98c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 17 20:42:30 2025 +0200

        tdf#130857 weld: Move signal block counter logic to weld::Widget

, now also move the logic to only conditionally call
the signal handlers from SalInstanceWidget (and subclasses)
to weld::Widget (and subclasses), so this can be reused
for other implementations in the future as well.

Don't add it this logic to all signal_* methods, but
only those for which the SalInstance* implementations
were doing this previously, to not change the existing
behavior.

The only known case where the behavior would slightly
differ now is SalInstanceCheckButton::ToggleHdl, which
previously was

    IMPL_LINK_NOARG(SalInstanceCheckButton, ToggleHdl, CheckBox&, void)
    {
        if (notify_events_disabled())
            return;
        m_xCheckButton->EnableTriState(false);
        signal_toggled();
    }

and now would still call

    m_xCheckButton->EnableTriState(false);

when event notification is disabled.
However, that should presumably be fine, and making
it dependent on whether or not signal notification
is enabled seems a little odd anyway.

The gtk3/gtk4 VCL plugin implement their own logic,
which remains unchanged for now.

Change-Id: I13bb1d4a2b52b2d4494a85895f2e33b98b2c45c6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192598
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>