tdf#130857 weld: Move signal blocking to TextView::select_region
So far, the vcl and gtk3/gtk4 implementations
have their separate implementations to avoid
signalling certain widget changes when those
changes are done programatically and not
manually by the user, and the native qt5/qt6 implementation
didn't yet implement any such logic, meaning
signals would be emitted more often for supported
dialogs in the SAL_VCL_QT_USE_WELDED_WIDGETS=1 case
than for the gtk and vcl implementations.

While emitting signals also on programmatic
changes seems plausible in general and matches
what e.g. the GTK and Qt toolkits do for their
widgets, it would cause various dialog implementations
to misbehave with the native Qt implementation as they
rely on the signals not being emitted in such cases.
See for example

    commit 3ffa3be35ee712db24ec9b2ffa55e96c3d97d169
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Aug 12 13:25:15 2025 +0200

        tdf#130857 cui: Refine toggle logic in "Area" tab page

for one commit to adjust a dialog to work as expected
when signals are emitted on programmatic changes as well.

In order to unify the behavior between the different
weld::Widget (and subclasses) implementations on when
signals are blocked without having to duplicate the
same logic in multiple places, calling the methods to block
and unblock signal emission is moved from the
vcl (SalInstance*) implementations to the weld::Widget
(and subclasses) methods.
In order to do so, introduce new helper methods to
implement in the toolkit-specific implementations
that implement the "actual" logic of modifying the
underlying toolkit widget, and are wrapped by
the weld::Widget method that first calls
weld::Widget::disable_notify_events to disable
signals before calling the "actual" logic, then
enables signals again by calling weld::Widget::enable_notify_events.
This is what the SalInstanceWidget implementation did
previously.

This means that the logic previously implemented
in the vcl implementation now also applies for
the qt5/qt6 and gtk3/gtk4 ones.

As described above, the native qt5/qt6 implementation
(QtInstanceWidget and subclasses) didn't have
any such mechanism yet, and now gets aligned with the
vcl implementation.

The native gtk3/gtk4 implementation (GtkInstanceWidget and
subclasses) already has its own logic to block signal emission
and now gets another level of filtering signals.
That should be fine, because only signals that are
not emitted for the vcl version are now filtered for gtk3/gtk4
on another level as well, so dialog implementations shouldn't
expect these signals.
After all, the existing gtk3/gtk4 logic to filter out
signals should be equivalent and presumably become
mostly redundant once the current vcl one is moved to
the weld::Widget base classes. It is left in place
for now, however, and can/should be revisited separately
on a case-by-case basis, to identify any logic that may
still be relevant.

This is the first commit in a series that applies the
logic change described above, and does so for
TextView::select_region.
The same will be done for the other methods in separate
upcoming commits.

See also the commit messages of previous commits

    Change-Id: I93ec71e29804fd601e314e24c96d0f18a7cbb98c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 17 20:42:30 2025 +0200

        tdf#130857 weld: Move signal block counter logic to weld::Widget

and

    Change-Id: I13bb1d4a2b52b2d4494a85895f2e33b98b2c45c6
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 17 21:49:10 2025 +0200

        tdf#130857 weld: Move logic for (not) calling signal hdls to base classes

for more background.

Change-Id: If5fba408cc6638441e9856238b19659fed6d7e01
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192610
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins