tdf#165472 sw: fix missing page cross-reference update on hidden sect change
Open the bugdoc, change the "book" set variable field's value at the
document start from 0 to 1, go to the end of the document, the page
cross-reference field says "page 3", but once you click on it, you end
up on page 2, which is not correct.

This is a regression from commit
1dcd3f1b9fc0d888e047e854a20e9ab2e0afd544 (sw: fix not needed
invalidation of custom field on each keypress, 2021-05-19), where the
old use-case was that typing in a LOK client resulted in repainting the
custom fields in the footer on each keystroke, because typing a key
changes the doc stats (one more character), but then by the time a
repaint happens, we don't know anymore what field types needs updating,
so we repaint all doc stat fields.

So repainting those fields on every keystroke is certainly an overkill,
but at the same time an automatic update of page cross-references on
section visibility change is also a reasonable request. Fix this by
separating the "update on idle after any cursor change" old use-case and
the "close dialog and enter text edit/standard mode" new use-case. Doing
this wider update of fields is enough to do in the second case, so we
get page number updates and the typing performance is not back to poor.

This works because the idle case goes via the vcl scheduler, but the
explicit field update goes via this codepath:
	#23 0x00007ffff2f088b1 in SfxBindings::Update (this=0x4c14b20, nId=21185) at sfx2/source/control/bindings.cxx:357
	#24 0x00007fffba614118 in SwWrtShell::Invalidate (this=0x4c5c980) at sw/source/uibase/wrtsh/select.cxx:66
	#25 0x00007fffba615fb9 in SwWrtShell::EnterStdMode (this=0x4c5c980) at sw/source/uibase/wrtsh/select.cxx:595
	#26 0x00007fff917e1af3 in SwFieldEditDlg::ImplDestroy (this=0x4c43a30) at sw/source/ui/fldui/fldedt.cxx:224
	#27 0x00007fff917e1b50 in SwFieldEditDlg::~SwFieldEditDlg (this=0x4c43a30, __in_chrg=<optimized out>) at sw/source/ui/fldui/fldedt.cxx:229

Change-Id: I96f48f46f44a1d6d1bde5ede282cd1cf847a6164
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192680
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>