FieldDescControl: Fix conditions needed to update the format sample
OnControlFocusLost has a long if-chain of actions to perform for each
control. The last one looks like it was trying to update the format
sample whenever the default value changes. However, there’s already a
case further up the chain that handles the case where the default value
text box loses focus and the value is different from the saved value.
That means the format sample would only end up being updated if the
default value control loses focus but nothing was changed in it, which
seems pretty clearly like a mistake. It makes it so the only way to make
the format sample update after changing the default value is to first
move the focus to another control, move it back to the default value and
then move it away again.

This patch just moves the format sample update into the same branch of
the if-chain that already exists to handle the default value control.

Change-Id: Ia4885589a7f50e01479972cceb4e7c6875efbde0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192753
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins