bin/find-unneeded-includes: another method against obsolete exclusions
Spotted with include/o3tl/vector_utils.hxx
it can happen that IWYU starts to consider a header having
all correct includes, in this case it does not give a list of
proposals.

This makes the previous method of spotting obsolete exclusions
ineffective in these cases.

Start watching for the "foo.hxx has correct #includes/fwd-decls"
announcement and consider every exclusion as obsolete

Change-Id: I46dd8dd4e2765a972ceaa7aa950d92c9e0a2d1b2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192747
Tested-by: Jenkins
Reviewed-by: Gabor Kelemen <gabor.kelemen@collabora.com>