Make the got/lose focus handlers use an iterator to access controls
Instead of having two long if-chains to handle the focus for the
controls, there is now an inline helper method that calls a function on
each of the controls. This is then used by the two handler methods to
check all of the controls. This should make it a bit easier to handle
the list of controls that need similar code.

Change-Id: I9787b92156dee39479a260d029cbd382fbef9394
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192975
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins