tdf#168970 OFieldDescControl: Add explicit method to flush changes
Previously when the table edit window was about to be closed it relied
on being able to trigger the OFieldDescControl to save its pending
changes by making it lose the keyboard focus. This doesn’t work on the
GTK3 backend and also seems like a bit of an unreliable approach anyway.
This patch instead adds a method to explicitly make the control flush
any pending changes.

Change-Id: I8e7b8e55ce4f56323fe1c0779bf3237ce847e5e8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192976
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins