Related: tdf#168916 DOCX import: handle page break from style on framed para
Open a subset of the bugdoc, it's of two pages in Word, it's one page in
Writer.

What happens is that the paragraph that's at the top of page 2 in Word
has a <w:framePr> property on it, so it forms a text frame, but it also
inherits a <w:pageBreakBefore> via using the Heading 1 style and Word
layout then handles this 'page break before' request.

Fix the problem by extending a previous fix in this area: commit
f6f53f76e15f5eecc5b6ce56e471c53cebfea8ad (tdf#123636 writerfilter:
handle deferred breaks on frames, 2019-04-30) already tried to handle
this, but it covered only the case when the page break property was set
directly. Now this is also found via a style.

The actual bugdoc needs more fixes, but this way at least we attempt to
split floating tables at places where the expected result of that split
is known, while previously we tried to split at the wrong positions, due
to the loss of the page break.

Change-Id: Icc24fcdb05a4375583917eb97b63f4a19dc4bd17
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192834
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit 1d833123014274a182ea42e92ca21e4ece736389)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192843
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>