Implement the database range behaviour by using similar
(but not the same) algorytm as ScFillMode::FILL, therefore
we can only expand a DBRange in one direction at a time
(same as in MSO). In case of an outside range the first Column
or first two ( because of header/total row) will be selected.
Also (will) simplify a bit the undo/redo parts.

Also use PointerStyle::SESize instead of PointerStyle::Cross.

Change-Id: Ia4ba85494f9b7824118f68277a59156e6ea2a036
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192839
Reviewed-by: Balazs Varga <balazs.varga@collabora.com>
Tested-by: Balazs Varga <balazs.varga@collabora.com>