tdf#168226 Fix rounding behavior in PtTo10Mu and TenMuToPt functions
When o3tl::convert is used with floating-point input, it provides double output
without rounding. Before the change to use o3tl::convert in commit
6c688ee8625f3f5e7be50912e02cdf59f6ae7e5f, the result was rounded.
After the mentioned commit, it truncated when converting double to int.

To match the previous behavior, the result of o3tl::convert must be rounded.
This restores the behavior that existed before switching to o3tl::convert.

Change-Id: Id54b1c74f320b6f4cfa49c13ce4182434e9e6a8e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192841
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>