tdf#168691 fix "Image Mode" and other toolbar buttons not working
commits
29bd8d14eb632e621bbafa1fbf65add12800dc77^..eaa6757c5cb7865435367f0b80f13a8170f125a8
added support for executing uno commands like .uno:GrafMode without
needing any UI interaction (earlier it relied on the UI to extract the
values), but it broke the existing toolbar buttons as the values from
the UI were not properly passed via the params.

This commit fixes that by passing the values extracted from UI
to the required params.

Signed-off-by: Mohit Marathe <mohit.marathe@collabora.com>
Change-Id: I9a572e3fab2ff37efb8b0e4fce8758c08e3ff8d9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192850
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Tested-by: Jenkins